BOT/requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# Environment
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Scheduler
APScheduler==3.10.4

# Google Calendar
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0

# HTTP
aiohttp==3.9.1
requests==2.31.0

# Redis (опционально)
redis==5.0.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Development
black==23.12.1
flake8==6.1.0
pytest==7.4.3
